Rename grammarfileparser.rb to grammar_file_parser.rb
...In order to follow generally-accepted conventions on the naming of Ruby
source files.
