+++ bin/racc2y
@@ -7,7 +7,7 @@
 # the GNU LGPL, Lesser General Public License version 2.1.
 # For details of the LGPL, see the file "COPYING".
 
-require 'racc/grammarfileparser'
+require 'racc/grammar_file_parser'
 require 'racc/info'
 require 'optparse'
